+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 * The Infusion Calculator has been removed, now that infusions are much more straightforward.
+* Pressing the "i" key on the keyboard will toggle showing item details in the item popup.
 
 # 3.5